refactor: componente de date-picker e integrar com period-picker
