+++ src/app/domain/events/containers/form-event/form-event.ts
@@ -11,10 +11,10 @@ import { MatTooltip } from '@angular/material/tooltip';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FormInputComponent } from '../../../../shared/components/form-input/form-input';
 import { ImageInputComponent } from '../../../../shared/components/image-input/image-input';
+import { DatePickerComponent } from '../../../../shared/components/date-picker/date-picker';
 import { CanComponentDeactivate } from '../../../../shared/guards/unsaved-changes.guard';
 import { FormHelperService } from '../../../../shared/services/form/form-helps';
 import { CKEditorComponent } from '../../components/ck-editor/ck-editor';
-import { DatePickerComponent } from '../../components/date-picker/date-picker';
 import { TimePickerComponent } from '../../components/time-picker/time-picker';
 import { EventService } from '../../services/event-service';
 import { DateTimeUtils } from '../../utils/date-time';
