+++ src/app/shared/components/date-picker/date-picker.html
@@ -4,6 +4,8 @@
     matInput
     [matDatepicker]="picker"
     [placeholder]="placeholder"
+    [min]="min"
+    [max]="max"
     [formControl]="formControl!"
     (input)="onInput($event)"
     (blur)="onBlur()"
