+++ src/app/shared/components/date-picker/date-picker.scss
